+++ js/utils.js
@@ -1,0 +1,9 @@
+// Функция-генератор id
+export function getGeneratedId() {
+  let lastId = 0;
+
+  return function () {
+    lastId += 1;
+    return lastId;
+  };
+}
